new validation to delete users

+++ src/main/java/br/com/macrosapi/services/UserService.java
@@ -4,6 +4,7 @@ import br.com.macrosapi.dto.RegisterUserDTO;
 import br.com.macrosapi.dto.UserDetailsDTO;
 import br.com.macrosapi.model.user.User;
 import br.com.macrosapi.repositories.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -18,6 +19,9 @@ public class UserService {
 
     @Autowired
     private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    private TokenService tokenService;
 
     public User create(RegisterUserDTO dto) {
         var encodedPassword = passwordEncoder.encode(dto.password());
@@ -34,8 +38,15 @@ public class UserService {
         throw new RuntimeException("This user was deleted");
     }
 
-    public void delete(UUID id) {
+    public void delete(UUID id, HttpServletRequest request) throws IllegalAccessException {
         User user = userRepository.getReferenceById(id);
+        String token = tokenService.recoverTokenFromCookies(request);
+        String email = tokenService.getSubject(token);
+
+        if (!email.equals(user.getEmail())) {
+            throw new IllegalAccessException("You can only delete your own account");
+        }
+
         user.delete();
     }
 }
